# documentacao/api.yml
openapi: "3.0.2"
info:
  title: Avaliação de Experiência
  version: "1.0"
servers:
  - url: localhost:3000
paths:
  /loja:
    post:
      description: Cria uma loja.
      requestBody:
        $ref: "#/components/requestBodies/PostLoja"
      responses:
        "200":
          $ref: "#/components/responses/Loja"
        "400":
          description: Loja não deve ser especificada.
        "404":
          description: Loja não encontrada.
        "500":
          description: Erro indefinido.
    put:
      description: Atualiza uma Loja.
      requestBody:
        $ref: "#/components/requestBodies/PutLoja"
      responses:
        "200":
          $ref: "#/components/responses/Loja"
        "500":
          description: Erro indefinido.
  /lojas:
    get:
      description: Lista Lojas com paginação.
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/pageSizeParam"
        - $ref: "#/components/parameters/orderByParam"
        - $ref: "#/components/parameters/orderDirectionParam"
        - $ref: "#/components/parameters/ativaParam"
      responses:
        "200":
          $ref: "#/components/responses/Lojas"
        "400":
          description: Parâmetro inválido.
        "500":
          description: Erro indefinido.
  /loja/{lojaId}:
    delete:
      description: Deleção lógica de uma Loja.
      parameters:
        - $ref: "#/components/parameters/lojaIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Loja"
        "400":
          description: Loja indefinida.
        "404":
          description: Loja não encontrada.
        "500":
          description: Erro indefinida.
  /colaborador:
    post:
      description: Cria um Colaborador.
      requestBody:
        $ref: "#/components/requestBodies/PostColaborador"
      responses:
        "200":
          $ref: "#/components/responses/Colaborador"
        "400":
          description: Colaborador não deve ser especificado.
        "404":
          description: Colaborador não encontrado.
        "500":
          description: Erro indefinido.
    put:
      description: Atualiza um Colaborador.
      requestBody:
        $ref: "#/components/requestBodies/PutColaborador"
      responses:
        "200":
          $ref: "#/components/responses/Colaborador"
        "500":
          description: Erro indefinido.
  /colaboradores:
    get:
      description: Lista Colaboradores com paginação.
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/pageSizeParam"
        - $ref: "#/components/parameters/orderByParam"
        - $ref: "#/components/parameters/orderDirectionParam"
        - $ref: "#/components/parameters/ativaParam"
      responses:
        "200":
          $ref: "#/components/responses/Colaboradores"
        "400":
          description: Parâmetro inválido.
        "500":
          description: Erro indefinido.
  /colaborador/{colaboradorId}:
    delete:
      description: Deleção lógica de um Colaborador.
      parameters:
        - $ref: "#/components/parameters/colaboradorIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Colaborador"
        "400":
          description: Colaborador indefinido.
        "404":
          description: Colaborador não encontrado.
        "500":
          description: Erro indefinida.
  /cliente:
    post:
      description: Cria um cliente.
      requestBody:
        $ref: "#/components/requestBodies/PostCliente"
      responses:
        "200":
          $ref: "#/components/responses/Cliente"
        "400":
          description: Cliente não deve ser especificado.
        "404":
          description: Cliente não encontrado.
        "500":
          description: Erro indefinido.
    put:
      description: Atualiza um cliente.
      requestBody:
        $ref: "#/components/requestBodies/PutCliente"
      responses:
        "200":
          $ref: "#/components/responses/Cliente"
        "500":
          description: Erro indefinido.
  /clientes:
    get:
      description: Lista clientes com paginação.
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/pageSizeParam"
        - $ref: "#/components/parameters/orderByParam"
        - $ref: "#/components/parameters/orderDirectionParam"
        - $ref: "#/components/parameters/ativoParam"
      responses:
        "200":
          $ref: "#/components/responses/Clientes"
        "400":
          description: Parâmetro inválido.
        "500":
          description: Erro indefinido.
  /cliente/{clienteId}:
    delete:
      description: Deleção lógica de um Cliente.
      parameters:
        - $ref: "#/components/parameters/clientIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Cliente"
        "400":
          description: Cliente indefinido.
        "404":
          description: Cliente não encontrado.
        "500":
          description: Erro indefinido.
  /transacao-experiencia:
    post:
      description: Cria uma Transação/Experiência.
      requestBody:
        $ref: "#/components/requestBodies/PostTransacaoExperiencia"
      responses:
        "200":
          $ref: "#/components/responses/TransacaoExperiencia"
        "400":
          description: Transação/Experiência não deve ser especificada ou Cliente indefinido.
        "404":
          description: Cliente inexistente ou  Loja inexistente ou Colaborador Inexistente.
        "500":
          description: Erro indefinido.
    put:
      description: Atualiza uma Transação/Experiência.
      requestBody:
        $ref: "#/components/requestBodies/PutTransacaoExperiencia"
      responses:
        "200":
          $ref: "#/components/responses/TransacaoExperiencia"
        "400":
          description: Transação/Experiência não especificada.
        "404":
          description: Transação/Experiência não encontrada.
        "500":
          description: Erro indefinido.
...
components:
  requestBodies:
    PostLoja:
      description: JSON contendo uma nova Loja.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              nome:
                type: string
                nullable: false
              ativa:
                type: boolean
                default: true
            example:
              nome: Loja 1
              ativa: true
    PutLoja:
      description: JSON contendo dados a serem salvos em um Loja.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
                nullable: false
              nome:
                type: string
                nullable: false
              ativa:
                type: boolean
                default: true
            example:
              id: 1
              nome: Loja A
              ativa: false
    PostColaborador:
      description: JSON contendo um novo Colaborador.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              nome:
                type: string
                nullable: false
              ativo:
                type: boolean
                default: true
            example:
              nome: Colaborador 1
              ativo: true
    PutColaborador:
      description: JSON contendo dados a serem salvos em um Colaborador.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
                nullable: false
              nome:
                type: string
                nullable: false
              ativa:
                type: boolean
                default: true
            example:
              id: 1
              nome: Colaborador A
              ativa: false
    PostCliente:
      description: JSON contendo um novo Cliente.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              nome:
                type: string
                nullable: false
              email:
                type: string
                nullable: false
              telefone:
                type: string
                nullable: false
              cpf:
                type: string
                nullable: false
              ativo:
                type: boolean
                default: true
            example:
              nome: João da Silva
              email: cliente@email.com
              telefone: 21986973989
              cpf: 05977788096
              ativo: true
    PutCliente:
      description: JSON contendo dados a serem salvos em um Cliente.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
                nullable: false
              nome:
                type: string
                nullable: false
              email:
                type: string
                nullable: false
              telefone:
                type: string
                nullable: false
              cpf:
                type: string
                nullable: false
              ativo:
                type: boolean
                default: true
            example:
              id: 1
              nome: João da Silva Jr
              email: cliente2@email.com
              telefone: 91987650174
              cpf: 23837266648
              ativo: false
    PostTransacaoExperiencia:
      description: JSON contendo dados da Transação\Experiência a ser criada.
      content:
        application/json:
          schema:
            type: object
            properties:
              valor:
                type: number
                nullable: false
              data:
                type: string
                nullable: false
              clienteId:
                type: integer
                nullable: false
              lojaId:
                type: integer
                nullable: false
            example:
              valor: 53.3
              data: "2021-06-20T03:00:00.000Z"
              clienteId: 2
              lojaId: 3
    PutTransacaoExperiencia:
      description: JSON contendo dados da Transação\Experiência a ser atualizada.
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
                nullable: false
              valor:
                type: number
                nullable: false
              data:
                type: string
                nullable: false
              clienteId:
                type: integer
                nullable: false
              lojaId:
                type: integer
                nullable: false
            example:
              id: 2
              valor: 53.3
              data: "2021-05-21T03:00:00.000Z"
              clienteId: 3
              lojaId: 1
...
  parameters:
    pageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: número da página para paginação.
    pageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: tamanho da página para paginação.
    orderByParam:
      name: ordeBy
      in: query
      required: false
      schema:
        type: string
        enum: [id, nome, email, telefone, cpf, creationDate, updateDate, ativo]
        example: nome
      description: coluna de ordenação.
    orderDirectionParam:
      name: orderDirectionParam
      in: query
      required: false
      schema:
        type: string
        enum: [ASC, DESC]
      description: direção de ordenação.
      example: ASC
    ativoParam:
      name: ativo
      in: query
      required: false
      schema:
        type: boolean
      description: filtrar registros ativos.
      example: true
    ativaParam:
      name: ativa
      in: query
      required: false
      schema:
        type: boolean
      description: filtrar registros ativos.
      example: true
    clientIdParam:
      name: clienteId
      in: path
      required: true
      schema:
        type: integer
      description: id do Cliente.
    lojaIdParam:
      name: lojaId
      in: path
      required: true
      schema:
        type: integer
      description: id do Loja.
    colaboradorIdParam:
      name: colaboradorId
      in: path
      required: true
      schema:
        type: integer
      description: id do Colaborador.
  responses:
    Loja:
      description: Loja salva.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Loja"
    Lojas:
      description: Lista de Lojas com paginação.
      content:
        application/json:
          schema:
            type: object
            properties:
              list:
                type: array
                items:
                  $ref: "#/components/schemas/Loja"
              order:
                type: object
                nullable: true
                example:
                  nome: "ASC"
              offset:
                type: integer
                minimum: 0
                example: 10
                description: número de itens antes do primeiro item da página.
              page:
                type: integer
                example: 2
                minimum: 1
                description: número da página de dados.
              pageSize:
                type: integer
                example: 10
                minimum: 1
                maximum: 100
                description: tamanho da página de dados.
              total:
                type: integer
                minimum: 0
                example: 120
                description: total de itens em todas as páginas.
    Colaborador:
      description: Colaborador salvo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Colaborador"
    Colaboradores:
      description: Lista de Colaboradores com paginação.
      content:
        application/json:
          schema:
            type: object
            properties:
              list:
                type: array
                items:
                  $ref: "#/components/schemas/Colaborador"
              order:
                type: object
                nullable: true
                example:
                  nome: "ASC"
              offset:
                type: integer
                minimum: 0
                example: 10
                description: número de itens antes do primeiro item da página.
              page:
                type: integer
                example: 2
                minimum: 1
                description: número da página de dados.
              pageSize:
                type: integer
                example: 10
                minimum: 1
                maximum: 100
                description: tamanho da página de dados.
              total:
                type: integer
                minimum: 0
                example: 120
                description: total de itens em todas as páginas.
    Cliente:
      description: Cliente salvo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Cliente"
    Clientes:
      description: Lista de Clientes com paginação.
      content:
        application/json:
          schema:
            type: object
            properties:
              list:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
              order:
                type: object
                nullable: true
                example:
                  nome: "ASC"
              offset:
                type: integer
                minimum: 0
                example: 10
                description: número de itens antes do primeiro item da página.
              page:
                type: integer
                example: 2
                minimum: 1
                description: número da página de dados.
              pageSize:
                type: integer
                example: 10
                minimum: 1
                maximum: 100
                description: tamanho da página de dados.
              total:
                type: integer
                minimum: 0
                example: 120
                description: total de itens em todas as páginas.
    TransacaoExperiencia:
      description: Transação/Experiência salva.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransacaoExperiencia"
...
  schemas:
    Loja:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        ativa:
          type: boolean
      example:
        id: 11
        nome: Loja 1
        creationDate: "2021-05-21T16:52:58.093Z"
        updateDate: "2021-05-21T16:52:58.093Z"
        ativa: true
    Colaborador:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        ativo:
          type: boolean
      example:
        id: 11
        nome: Paulo Henrique
        creationDate: "2021-05-21T16:52:58.093Z"
        updateDate: "2021-05-21T16:52:58.093Z"
        ativo: true
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cpf:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        ativo:
          type: boolean
      example:
        id: 11
        nome: João da Silva
        email: joaodasilva@email.com
        cpf: "62192726090"
        telefone: "27996697580"
        ativo: true
        creationDate: "2021-05-21T16:52:58.093Z"
        updateDate: "2021-05-21T16:52:58.093Z"
    TransacaoExperiencia:
      type: object
      properties:
        id:
          type: integer
        valor:
          type: number
        data:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        clienteId:
          type: integer
        lojaId:
          type: integer
      example:
        id: 5
        valor: 53.23
        data: "2020-05-21T00:00:00.000Z"
        creationDate: "2021-05-21T16:52:58.093Z"
        updateDate: "2021-05-21T16:52:58.093Z"
        clienteId: 7
        lojaId: 3
...